+++ clients/simple-earn/src/main/java/com/binance/connector/client/simple_earn/rest/api/RwusdApi.java
@@ -53,7 +53,7 @@ public class RwusdApi {
 
     private static final String USER_AGENT =
             String.format(
-                    "binance-simple-earn/5.0.0 (Java/%s; %s; %s)",
+                    "binance-simple-earn/6.0.0 (Java/%s; %s; %s)",
                     SystemUtil.getJavaVersion(), SystemUtil.getOs(), SystemUtil.getArch());
     private static final boolean HAS_TIME_UNIT = false;