+++ examples/spot/src/main/java/com/binance/connector/client/spot/rest/trade/OrderCancelReplaceExample.java
@@ -43,10 +43,12 @@ public class OrderCancelReplaceExample {
     /**
      * Cancel an Existing Order and Send a New Order
      *
-     * <p>Cancels an existing order and places a new order on the same symbol. Filters and Order
-     * Count are evaluated before the processing of the cancellation and order placement occurs. A
+     * <p>* Cancels an existing order and places a new order on the same symbol. * Filters and Order
+     * Count are evaluated before the processing of the cancellation and order placement occurs. * A
      * new order that was not attempted (i.e. when &#x60;newOrderResult: NOT_ATTEMPTED&#x60;), will
-     * still increase the unfilled order count by 1. Weight: 1
+     * still increase the unfilled order count by 1. * You can only cancel an individual order from
+     * an orderList using this endpoint, but the result is the same as canceling the entire
+     * orderList. Weight: 1
      *
      * @throws ApiException if the Api call fails
      */